% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikzsetup.R
\name{tikzsetup}
\alias{tikzsetup}
\title{Setup tikz device for russian Rmd/Rnw files}
\usage{
tikzsetup(compiler = "xetex", lang = "russian", cyrillic_font = "Arial",
  font_size = 12)
}
\arguments{
\item{compiler}{currently xelatex is supported}

\item{lang}{language}

\item{cyrillic_font}{Arial by default}

\item{font_size}{font size}
}
\value{
invisible preamble
}
\description{
Setup tikz device for russian Rmd/Rnw files
}
\details{
Setup tikz device for russian Rmd/Rnw files
}
\examples{
tikzsetup()
tikzsetup(lang = "russian")
}
